Move libraries to their own folders
